[BCO-4773] Support only integers for sortValue

Closes BCO-4773

See merge request aboutyou/cc-babo/admin-api!200

+++ src/main/java/com/aboutyou/cloud/adminapi/model/ProductSorting.java
@@ -39,6 +39,6 @@ public class ProductSorting implements ApiObjectInterface {
     * The value used for sorting.
     */
     @SerializedName("sortValue")
-    Double sortValue;
+    Integer sortValue;
 
 }
